[#184] 펀딩 서비스 이벤트 통신 리팩토링 (#185)
* chore: 테스트 코드 임포트문 최적화

* refactor: 이벤트 발행 코드 최적화

* chore: funding-service 임포트문 최적화

* test: funding create test 수정

* refactor: 아웃박스 테이블에 페이로드 타입 포함

카프카 발행 전 어떤 타입의 페이로드인지 확인하고 매핑하기 위해 추가

* test: 이벤트 코드 리팩토링으로 테스트 코드 재작성

* refactor: 프로젝트 종료 처리 로직을 성공/실패 케이스별로 분리

- processSuccessfulProject() 메서드 추출
- processFailedProject() 메서드 추출
- closeProcessingProjects()의 가독성 개선

+++ funding/src/main/java/com/nowayback/funding/application/fundingProjectStatistics/dto/event/ProjectFundingFailedEvent.java
@@ -37,7 +37,7 @@ public record ProjectFundingFailedEvent(
     public static ProjectFundingFailedEvent of(
             UUID projectId,
             Long finalAmount,
-            Long participantCount,
+            Integer participantCount,
             Long targetAmount,
             Double achievementRate
     ) {
