+++ funding/src/main/java/com/nowayback/funding/application/fundingProjectStatistics/dto/event/payload/ProjectFundingSuccessPayload.java
@@ -5,5 +5,5 @@ import java.util.UUID;
 public record ProjectFundingSuccessPayload(
         UUID projectId,
         Long finalAmount,
-        Long participantCount
+        Integer participantCount
 ) {}
